+++ src/transport/xmpp/xmpp_xml.erl
@@ -10,9 +10,27 @@
          create_session/0, 
          bind/1, 
          muc/1, 
-         message/3]).
+         message/2,
+         private_message/2,
+         presence/0,
+         presence_subscribed/1]).
 
 -include("xmpp.hrl").
+
+%% @doc send online status to all
+presence() ->
+    % xml data structure
+    XmlData = [{'presence', [{'priority', '50'}, {'show', 'chat'}, {'status', 'Hello, from Ybot'}], []}],
+    % convert to xml
+    lists:flatten(xmerl:export_simple(XmlData, xmerl_xml, [{prolog, ""}])).
+
+%% @doc send 'subscribed' presence
+-spec presence_subscribed(To :: string()) -> string().
+presence_subscribed(To) ->
+    % xml data structure
+    XmlData = [{'presence', [{'to', To}, {'type', 'subscribed'}], []}],
+    % convert to xml
+    lists:flatten(xmerl:export_simple(XmlData, xmerl_xml, [{prolog, ""}])).
 
 %% @doc xmpp plain authentication
 -spec auth_plain(Login :: string()) -> string().
@@ -51,11 +69,22 @@ create_session() ->
     % convert to xml
     lists:flatten([xmerl:export_simple(XmlData, xmerl_xml, [{prolog, ""}])]).
 
+%% @doc send private message
+-spec private_message(To :: string(), Message :: string()) -> string().
+private_message(To, Message) ->
+  % xml data structure
+  XmlData = [{'message', [{'to', To},  {'type', "chat"}],
+              [{'body', [Message]}]
+            }],
+  
+  % convert to xml
+  lists:flatten(xmerl:export_simple(XmlData, xmerl_xml, [{prolog, ""}])).
+
 %% @doc Send message to muc
--spec message(From :: string(), Room :: string(), Message :: string()) -> string().
-message(From, Room, Message) ->
+-spec message(Room :: string(), Message :: string()) -> string().
+message(Room, Message) ->
     % xml structure data
-    XmlData = [{'message', [{'to', Room}, {'from', From}, {'type', "groupchat"}],
+    XmlData = [{'message', [{'to', Room}, {'type', "groupchat"}],
                 [{'body', [Message]}]
               }],
     % convert to xml