+++ Changelog.md
@@ -19,6 +19,13 @@
   * Added http transport.
   * Added PASS paraemetr to IRC.
   * Irc private messages support added --> # 24
+  * XMPP-muc private message supporting added.
+  * XMPP single user chat supporting added.
+  * XMPP ssl support added.
+  * GTALK support added.
+  * New option xmpp port added. 
+  * New option xmpp use_ssl added.
+  * Transport options validating added.
 
 ## Ybot-0.2 --> released